HandTranslatorASLAlphabet1/requirements.txt
opencv-python
mediapipe
torch
torchvision
numpy
pillow
tqdm
scikit-learn
flask-cors
